Styled mobile menu and fixed conditional header rendering
