+++ src/App.tsx
@@ -1,24 +1,23 @@
 import React from "react";
 import { Switch, Route, HashRouter } from "react-router-dom";
-import { Box, ThemeProvider } from "@material-ui/core";
+import {Box, ThemeProvider} from "@material-ui/core";
 import { theme } from "./theme";
-import { About } from "./pages/About";
-import { Home } from "./pages/Home";
-import { Careers } from "./pages/Careers";
 import { routes } from "./constants/routes";
-import { Header } from "./components/Header";
-import { Footer } from "./components/Footer";
-
-import "./App.css";
+import { Home } from "./pages/Home";
+import { About } from "./pages/About";
+import { Careers } from "./pages/Careers";
 import {Contact} from "./pages/Contact";
-import {HeaderMobile} from "./components/HeaderMobile";
+import {Header} from "./components/Header";
+import { Footer } from "./components/Footer";
+import "./App.css";
 
 const App: React.FC = () => {
+
   return (
     <ThemeProvider theme={theme}>
       <Box bgcolor="primary.main" width="100%" minHeight='100vh'>
         <HashRouter>
-          <HeaderMobile />
+          <Header/>
           <Switch >
             <Route path={routes.home.path} exact>
               <Home onClick={() => console.log("hey")} />
